Graph's components are refactored to have their own file
