+++ js/components/draw/main.js
@@ -1,4 +1,4 @@
-import { renderReactGraph } from '../graph/graph.js.jsx';
+import renderReactGraph from '../graph/renderReactGraph';
 
 document.addEventListener('DOMContentLoaded', () => {
     renderReactGraph('react-graph', true, true);
